refactor(core): move M10 runtime decisions into agent-runtime
Move provider-neutral DeepReview task execution shaping, DeepResearch citation renumbering, and prompt environment facts into bitfun-agent-runtime while keeping core as the compatibility and IO adapter boundary.

Update boundary checks and execution plans to separate completed provider-neutral M10 work from remaining concrete workflow and side-effect migrations.
